chore(rpc): move trace_at function to trait
